Replace UNHEX with CAST(UNHEX())
