+++ pkg/sql/plan/function/func_unary.go
@@ -16,6 +16,7 @@ package function
 
 import (
 	"context"
+	"encoding/base64"
 	"encoding/hex"
 	"fmt"
 	"io"
@@ -837,6 +838,25 @@ func Unhex(parameters []*vector.Vector, result vector.FunctionResultWrapper, pro
 	}
 
 	return nil
+}
+
+func ToBase64(ivecs []*vector.Vector, result vector.FunctionResultWrapper, proc *process.Process, length int) (err error) {
+	return opUnaryBytesToBytesWithErrorCheck(ivecs, result, proc, length, func(data []byte) ([]byte, error) {
+		buf := make([]byte, base64.StdEncoding.EncodedLen(len(functionUtil.QuickBytesToStr(data))))
+		base64.StdEncoding.Encode(buf, data)
+		return buf, nil
+	})
+}
+
+func FromBase64(ivecs []*vector.Vector, result vector.FunctionResultWrapper, proc *process.Process, length int) (err error) {
+	return opUnaryBytesToBytesWithErrorCheck(ivecs, result, proc, length, func(data []byte) ([]byte, error) {
+		buf := make([]byte, base64.StdEncoding.DecodedLen(len(functionUtil.QuickBytesToStr(data))))
+		_, err := base64.StdEncoding.Decode(buf, data)
+		if err != nil {
+			return nil, err
+		}
+		return buf, nil
+	})
 }
 
 func Length(ivecs []*vector.Vector, result vector.FunctionResultWrapper, proc *process.Process, length int) error {
