+++ test/distributed/cases/array/array.sql
@@ -76,14 +76,14 @@ select * from t4;
 
 -- insert vector as binary
 create table t5(a int, b vecf32(3));
-insert into t5 values(1, unhex('7e98b23e9e10383b2f41133f'));
-insert into t5 values(2, unhex('0363733ff13e0b3f7aa39d3e'));
-insert into t5 values(3, unhex('be1ac03e485d083ef6bc723f'));
+insert into t5 values(1, cast(unhex('7e98b23e9e10383b2f41133f') as blob));
+insert into t5 values(2, cast(unhex('0363733ff13e0b3f7aa39d3e') as blob));
+insert into t5 values(3, cast(unhex('be1ac03e485d083ef6bc723f') as blob));
 
 insert into t5 values(4, "[0,2,3]");
 
-insert into t5 values(5, unhex('05486c3f3ee2863e713d503dd58e8e3e7b88743f')); -- this is float32[5]
-insert into t5 values(6, unhex('9be2123fcf92de3e')); -- this is float32[2]
+insert into t5 values(5, cast(unhex('05486c3f3ee2863e713d503dd58e8e3e7b88743f') as blob)); -- this is float32[5]
+insert into t5 values(6, cast(unhex('9be2123fcf92de3e') as blob)); -- this is float32[2]
 
 select * from t5;
 select * from t5 where t5.b > "[0,0,0]";
